pyproject.toml
[project]
name = "smartdoc-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "datasets>=4.4.1",
    "flashrank>=0.2.10",
    "langchain>=1.0.8",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-google-vertexai>=3.0.3",
    "langchain-huggingface>=1.0.1",
    "pydantic>=2.11.10",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.0",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
    "unstructured[all-docs]>=0.18.20",
]
